+++ uninstall.ps1
@@ -12,6 +12,12 @@ param(
 
 # Based on @nerdio01's version in https://github.com/microsoft/terminal/issues/1060
 
+if ((Test-Path "Registry::HKEY_CLASSES_ROOT\Directory\shell\MenuTerminal") -and
+    -not (Test-Path "Registry::HKEY_CURRENT_USER\SOFTWARE\Classes\Directory\shell\MenuTerminal")) {
+    Write-Error "Please execute uninstall.old.ps1 to remove previous installation."
+    exit 1
+}
+
 $localCache = "$Env:LOCALAPPDATA\Microsoft\WindowsApps\Cache"
 if (Test-Path $localCache) {
     Remove-Item $localCache -Recurse
